Create a main Kyma Graph
Main Kyma graph will have all the workflow of the companion. It'll have init node too.
