+++ src/agents/common/data.py
@@ -4,4 +4,7 @@ from pydantic import BaseModel
 class Message(BaseModel):
     """Message data model."""
 
-    question: str
+    query: str
+    resource_type: str | None
+    resource_name: str | None
+    namespace: str | None
