+++ src/agents/kyma/agent.py
@@ -1,16 +1,14 @@
 import functools
+from typing import Any
 
 from langchain_core.tools import tool
 
+from agents.common.state import AgentState
 from agents.common.utils import agent_node, create_agent
 from utils.logging import get_logger
-from utils.models import LLM, ModelFactory
+from utils.models import Model
 
 logger = get_logger(__name__)
-
-model_factory = ModelFactory()
-model = model_factory.create_model(LLM.GPT4O_MODEL)
-
 KYMA_AGENT = "KymaAgent"
 
 
@@ -25,12 +23,23 @@ def search_kyma_doc(query: str) -> str:
     )
 
 
-kyma_agent_node = functools.partial(
-    agent_node,
-    agent=create_agent(
-        model.llm,
-        [search_kyma_doc],
-        "You are Kyma expert. You assist users with Kyma related questions.",
-    ),
-    name=KYMA_AGENT,
-)
+class KymaAgent:
+    """Supervisor agent class."""
+
+    name: str = KYMA_AGENT
+
+    def __init__(self, model: Model):
+        self.kyma_agent_node = functools.partial(
+            agent_node,
+            agent=create_agent(
+                model.llm,
+                [search_kyma_doc],
+                "You are Kyma expert. You assist users with Kyma related questions.",
+            ),
+            name=KYMA_AGENT,
+        )
+
+    def agent_node(self, state: AgentState) -> dict[str, Any]:
+        """Agent node."""
+
+        return self.kyma_agent_node(state)
