+++ src/services/messages.py
@@ -1,7 +1,9 @@
 import os
+from collections.abc import AsyncIterator
 
+from agents.common.data import Message
+from agents.graph import Graph, KymaGraph
 from agents.memory.redis_checkpointer import RedisSaver, initialize_async_pool
-from agents.supervisor.agent import Message, SupervisorAgent
 from utils.logging import get_logger
 from utils.models import LLM, ModelFactory
 
@@ -11,7 +13,7 @@ logger = get_logger(__name__)
 class MessagesService:
     """Chat service."""
 
-    supervisor_agent: SupervisorAgent
+    kyma_graph: Graph
     model_factory = ModelFactory()
 
     def __init__(self):
@@ -19,17 +21,19 @@ class MessagesService:
         memory = RedisSaver(
             async_connection=initialize_async_pool(url=f"{os.getenv('REDIS_URL')}/0")
         )
-        self.supervisor_agent = SupervisorAgent(model, memory)
+        self.kyma_graph = KymaGraph(model, memory)
 
     async def init_chat(self) -> dict:
         """Initialize the chat"""
         logger.info("Initializing chat...")
         return {"message": "Chat is initialized!"}
 
-    async def handle_request(self, conversation_id: int, message: Message):
+    async def handle_request(
+        self, conversation_id: int, message: Message
+    ) -> AsyncIterator[bytes]:
         """Handle a request"""
         logger.info("Processing request...")
 
-        async for chunk in self.supervisor_agent.astream(conversation_id, message):
+        async for chunk in self.kyma_graph.astream(conversation_id, message):
             logger.debug(f"Sending chunk: {chunk}")
             yield f"{chunk}\n\n".encode()
