+++ src/utils/langfuse.py
@@ -1,0 +1,9 @@
+import os
+
+from langfuse.callback import CallbackHandler
+
+handler = CallbackHandler(
+    secret_key=os.getenv("LANGFUSE_SECRET_KEY"),
+    public_key=os.getenv("LANGFUSE_PUBLIC_KEY"),
+    host=os.getenv("LANGFUSE_HOST"),
+)
