+++ tests/services/test_chat.py
@@ -23,7 +23,7 @@ class TestChat:
     def chat_instance(self, mock_supervisor_agent):
         with patch("services.chat.Chat.__init__", return_value=None):
             chat_object = MessagesService()
-            chat_object.supervisor_agent = mock_supervisor_agent
+            chat_object.kyma_graph = mock_supervisor_agent
             yield chat_object
 
     @pytest.mark.asyncio
@@ -33,7 +33,7 @@ class TestChat:
 
     @pytest.mark.asyncio
     async def test_handle_request(self, chat_instance, mock_supervisor_agent):
-        message = Message(question="Test message")
+        message = Message(query="Test message")
         result = [chunk async for chunk in chat_instance.handle_request(message)]
         mock_supervisor_agent.astream.assert_called_once_with(message)
         assert result == [b"chunk1\n\n", b"chunk2\n\n"]
